requirements.txt
streamlit>=1.25.0
psycopg2-binary>=2.9.5
pandas>=1.5.3
matplotlib>=3.7.1
seaborn>=0.12.2
numpy>=1.24.2
python-dotenv>=1.0.0 